requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
prometheus-fastapi-instrumentator==8.0.0
opentelemetry-api==1.9.1
opentelemetry-sdk==1.9.1
opentelemetry-instrumentation-fastapi==0.34b0
opentelemetry-exporter-otlp==1.9.1
requests==2.28.2
